requirements.txt
opencv-python
open3d
numpy
pandas
scipy
sklearn
tensorflow-gpu
